# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SmartStore SaaS API
  description: |
    A comprehensive REST API for managing e-commerce operations including products, orders, customers, and analytics.
    
    ## Authentication
    All API endpoints require authentication using NextAuth.js with JWT tokens.
    
    ## Rate Limiting
    API endpoints are rate limited to prevent abuse:
    - Authentication endpoints: 10 requests per minute
    - CRUD operations: 100 requests per 15 minutes
    - Analytics endpoints: 50 requests per 15 minutes
    - Monitoring endpoints: 200 requests per 15 minutes
    
    ## Support
    For API support and questions:
    - Documentation: https://docs.smartstore-saas.com
    - Support Email: support@smartstore-saas.com
    - Status Page: https://status.smartstore-saas.com
  version: 1.2.0
  contact:
    name: SmartStore SaaS Support
    email: support@smartstore-saas.com
    url: https://docs.smartstore-saas.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://smart-store-saas-demo.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get Products
      description: Retrieve a paginated list of products with optional filtering and sorting
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search term for name, description, or SKU
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category ID
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, price, createdAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  message:
                    type: string
                    example: Products fetched successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Products
      summary: Create Product
      description: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Product'
                  message:
                    type: string
                    example: Product created successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Product with SKU already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Products
      summary: Update Product
      description: Update an existing product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Product'
                  message:
                    type: string
                    example: Product updated successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Products
      summary: Delete Product
      description: Delete a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  example: product-123
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /security/audit:
    get:
      tags:
        - Security
      summary: Security Audit
      description: Perform security audit and get security recommendations
      responses:
        '200':
          description: Security audit completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SecurityAudit'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: product-123
        name:
          type: string
          example: Sample Product
        description:
          type: string
          example: Product description
        sku:
          type: string
          example: SKU-001
        price:
          type: number
          format: decimal
          example: 29.99
        cost:
          type: number
          format: decimal
          example: 15.00
        stock:
          type: integer
          example: 100
        isActive:
          type: boolean
          example: true
        categoryId:
          type: string
          example: cat-123
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
    ProductCreate:
      type: object
      required:
        - name
        - sku
        - price
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: New Product
        description:
          type: string
          maxLength: 1000
          example: Product description
        sku:
          type: string
          minLength: 1
          maxLength: 100
          example: SKU-002
        price:
          type: number
          minimum: 0.01
          example: 39.99
        cost:
          type: number
          minimum: 0
          example: 20.00
        categoryId:
          type: string
          example: cat-123
        isActive:
          type: boolean
          default: true
    ProductUpdate:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          example: product-123
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        sku:
          type: string
          minLength: 1
          maxLength: 100
        price:
          type: number
          minimum: 0.01
        cost:
          type: number
          minimum: 0
        categoryId:
          type: string
        isActive:
          type: boolean
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
...
    SecurityAudit:
      type: object
      properties:
        environment:
          type: object
          properties:
            secure:
              type: boolean
              example: true
            issues:
              type: array
              items:
                type: string
        headers:
          type: object
          properties:
            https:
              type: boolean
              example: true
            userAgent:
              type: string
              example: curl/8.7.1
        recommendations:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error type
        message:
          type: string
          example: Human-readable error message
        code:
          type: string
          example: ERROR_CODE
        timestamp:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        path:
          type: string
          example: /api/endpoint
        requestId:
          type: string
          example: req_1234567890_abcdef
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email is required
              code:
                type: string
                example: required
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...